+++ crates/db-storage/src/lib.rs
@@ -39,5 +39,6 @@ pub mod sql_types {
     pub use super::events::EventExceptionKindType as Event_exception_kind;
     pub use super::events::EventInviteStatusType as Event_invite_status;
     pub use diesel::sql_types::*;
+    pub use types::core::InviteRoleType as Invite_role;
     pub use types::core::TariffStatusType as Tariff_status;
 }